Move EnterPasswordPage container logic to CustomSignUp container
